// lib/api/generated/model/sentenceClozeDataDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API Documentation
 * The API description
 * OpenAPI spec version: 1.0
 */
import type { SentenceClozeDataDtoPartsItem } from './sentenceClozeDataDtoPartsItem';
import type { TokenDto } from './tokenDto';
import type { SentenceClozeDataDtoAnswerBySlot } from './sentenceClozeDataDtoAnswerBySlot';

export interface SentenceClozeDataDto {
  promptKorean?: string;
  parts: SentenceClozeDataDtoPartsItem[];
  choices: TokenDto[];
  answerBySlot: SentenceClozeDataDtoAnswerBySlot;
  slotIds: string[];
}